# lib/api-zod/src/generated/types/auditLog.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the nursery management platform
 * OpenAPI spec version: 0.1.0
 */
import type { AuditLogAfter } from './auditLogAfter';
import type { AuditLogBefore } from './auditLogBefore';

export interface AuditLog {
  id: number;
  actorId: string;
  /** @nullable */
  actorRole?: string | null;
  operation: string;
  entityType: string;
  /** @nullable */
  entityId?: string | null;
  /** @nullable */
  before?: AuditLogBefore;
  /** @nullable */
  after?: AuditLogAfter;
  createdAt: string;
}